--- src/components/Icons.tsx
@@ -1,44 +1,0 @@
-import { TbBrandVscode, TbBrandCypress } from "react-icons/tb";
-import {
-  SiCss3,
-  SiJavascript,
-  SiTypescript,
-  SiReact,
-  SiRedux,
-  SiNextdotjs,
-  SiNodedotjs,
-  SiBootstrap,
-  SiSass,
-  SiTailwindcss,
-  SiJest,
-  SiVuedotjs,
-  SiAngularjs,
-  SiFigma,
-  SiGit,
-  SiGithub,
-  SiHtml5,
-  SiChakraui
-} from "react-icons/si";
-
-export const icons = [
-    { icon: <SiHtml5 />, color: "#F1662A", delay: 0.6 },
-    { icon: <SiCss3 />, color: "#33A9DC", delay: 0.7 },
-    { icon: <SiJavascript />, color: "#F5DE19", delay: 0.8 },
-    { icon: <SiTypescript />, color: "#0065A9", delay: 0.9 },
-    { icon: <SiReact />, color: "#00D8FF", delay: 1 },
-    { icon: <SiRedux />, color: "#764ABC", delay: 1.1 },
-    { icon: <SiVuedotjs />, color: "#68a063", delay: 1.2 },
-    { icon: <SiAngularjs />, color: "#dd1b16", delay: 1.3 },
-    { icon: <SiNextdotjs />, color: "inherit", delay: 1.4 },
-    { icon: <SiNodedotjs />, color: "#68a063", delay: 1.5 },
-    { icon: <SiBootstrap />, color: "#563D7C", delay: 1.6 },
-    { icon: <SiTailwindcss />, color: "#44A8B3", delay: 1.7 },
-    { icon: <SiSass />, color: "#CD6799", delay: 1.8 },
-    { icon: <SiGit />, color: "#DE4C36", delay: 1.9 },
-    { icon: <SiGithub />, color: "inherit", delay: 2.0 },
-    { icon: <TbBrandVscode />, color: "#007ACC", delay: 2.1 },
-    { icon: <TbBrandCypress />, color: "inherit", delay: 2.22 },
-    { icon: <SiJest />, color: "#a6120d", delay: 2.3 },
-    { icon: <SiFigma />, color: "inherit", delay: 2.4 },
-    { icon: <SiChakraui />, color: "#03fcca", delay: 2.5 },
-  ];
