+++ src/data/link.ts
@@ -1,0 +1,7 @@
+export const link: { link: string; href: string; delay: number }[] = [
+  { link: "Home", href: "#Home", delay: 0.1 },
+  { link: "Projects", href: "#Projects", delay: 0.3 },
+  { link: "Tech Stacks", href: "#TechStack", delay: 0.4 },
+  { link: "About", href: "#About", delay: 0.2 },
+  { link: "Contacts", href: "#Contacts", delay: 0.5 },
+];
