+++ src/data/theme.tsx
@@ -1,0 +1,21 @@
+export const lightTheme = {
+  primaryReverse: "#A7A7A7",
+  primary: "#666666",
+  secondary: "#42446E",
+  standard: "#ffffff",
+  border: "#e7ebf0",
+  card: "#000",
+  cardBG: "#fff",
+  cardBS: "rgba(0, 0, 0, 0.2)",
+};
+
+export const darkTheme = {
+  primaryReverse: "#666666",
+  primary: "#A7A7A7",
+  secondary: "#CCCCCC",
+  standard: "#191919",
+  border: "#212020",
+  card: "#fff",
+  cardBG: "#363636",
+  cardBS: "rgba(0, 0, 0, 0)",
+};
